requirements.txt
PyQt6>=6.4.0
PyInstaller>=5.0.0
Pillow>=9.0.0
pywin32>=306
pypdf>=4.0.0
PyMuPDF>=1.23.0
